requirements.txt
shioaji>=1.2.6
python-dotenv>=1.0.1
pandas>=2.2.0
